refactor : separate homepage based on functions
