+++ src/features/menu/MenuItem.jsx
@@ -1,0 +1,5 @@
+import styled from '@emotion/styled';
+
+export default function MenuItem() {
+  return <div>MenuItem</div>;
+}
