+++ src/features/menu/index.js
@@ -1,0 +1,3 @@
+export { default as MenuListView } from './MenuListView';
+export { default as MenuGridView } from './MenuGridView';
+export { default as MenuTimeSelectorButton } from './MenuTimeSelectorButton';
