+++ src/features/review/components/index.js
@@ -1,4 +1,3 @@
 export { default as ReviewInput } from './ReviewInput';
-export { default as ReviewItem } from './ReviewItem';
-export { default as ReviewPagination } from './ReviewPagination';
+export { default as ReviewScrollView } from './ReviewScrollView';
 export { default as ReviewStarRatingSelector } from './ReviewStarRatingSelector';
