+++ src/features/review/index.js
@@ -1,4 +1,1 @@
 export * from './components';
-export * from './constants';
-export * from './datas';
-export * from './hooks';
