+++ src/shared/stores/index.js
@@ -1,1 +1,2 @@
 export { useSelectedRestaurantStore } from './useSelectedRestaurantStore';
+export { useSelectedDormMenuStore } from './useSelectedDormMenuStore';
