+++ src/shared/stores/useSelectedDormMenuStore.js
@@ -1,0 +1,12 @@
+import { create } from 'zustand';
+
+export const useSelectedDormMenuStore = create((set) => ({
+  clicked: 1,
+  setClicked: (id) => set(() => ({ clicked: id })),
+  selectedMenu: null,
+  setSelectedMenu: (menu) => set(() => ({ selectedMenu: menu })),
+  weeklyMenu: null,
+  setWeeklyMenu: (menu) => set(() => ({ weeklyMenu: menu })),
+  todayMenu: null,
+  setTodayMenu: (menu) => set(() => ({ todayMenu: menu })),
+}));
